+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Increase "dead-worker" detection to 20 minutes of unchanged beat
 * Reduce running task checking interval to every 5min after first check
 * Fix workeradmin tasks taking all bandwidth
+* Improve management of CPU-tasks
 
 ## [v1.0.0](https://github.com/OpenNMT/nmt-wizard/releases/tag/v1.0.0) (2018-10-04)